Initial version
RMG_utils.py -- these utility functions are used by the Ready–Made Garments (RMG) Observation script(s).  These functions are not specific to RMG use cases but instead are general enough to be useful in other contexts.   We have therefore placed them in this separate file to increase the likelihood of reuse.  This approach should reduce both the size and the complexity of RMG_specific scripts.